section_two/blog_project/blog_app/serializers.py
from rest_framework import generics, serializers
from .models import Blog, Comment

class BlogSerializer(serializers.ModelSerializer):
    class Meta:
        model = Blog
        fields = ['id', 'title', 'content', 'author', 'date_posted', 'date_updated', 'thumbnail', 'comment_count']
        
class CommentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Comment
        fields = ['id', 'blog', 'author', 'body', 'date_commented']

